refactor: 💡自定義工具整理至 common/utils
