+++ src/common/utils/apiExample/index.ts
@@ -1,0 +1,1 @@
+export * from './swaggerBasicExample';
