+++ src/common/utils/apiExample/swaggerBasicExample.ts
@@ -1,0 +1,4 @@
+export const basicExample = {
+  status: 'success',
+  message: '成功',
+};
